requirements.txt
pretrainedmodels==0.7.4
torch==2.5.1
efficientnet_pytorch==0.7.1
matplotlib
torchvision==0.20.1
numpy==1.26.4
pillow==11.0.0
joblib==1.4.2
pandas==2.2.3
natsort==8.4.0
opencv-python-headless==4.10.0.84
albumentations==1.3.1
scikit-learn==1.5.2
pydicom 
eel == 0.18.1
timm==0.6.13
catboost
